git-svn-id: http://svn.beutner.name/trunk/sbnc@926 b3a6972a-4407-0410-a3f0-9bc6c1699d69

+++ src/rpc/RPCApi.cpp
@@ -58,7 +58,7 @@ int GetStdHandle(int Handle) {
 	return Handle;
 }
 
-int ReadFile(int File, void *Buffer, size_t Size, int *Read, void *Dummy) {
+int ReadFile(int File, void *Buffer, int Size, int *Read, void *Dummy) {
 	int Result;
 
 	if (Size == 0) {
@@ -86,7 +86,7 @@ int ReadFile(int File, void *Buffer, size_t Size, int *Read, void *Dummy) {
 	}
 }
 
-int WriteFile(int File, const void *Buffer, size_t Size, int *Written, void *Dummy) {
+int WriteFile(int File, const void *Buffer, int Size, int *Written, void *Dummy) {
 	size_t Offset = 0;
 	int Result;
